requirements.txt
Flask-SQLAlchemy~=3.1.1
Flask~=2.3.3
mysql-connector-python~=8.3.0
python-dotenv~=1.0.0
flask-restx~=1.2.0
PyJWT~=2.8.0
numpy~=1.26.2
requests~=2.32.3
Flask-Cors~=4.0.1
ttkbootstrap~=1.10.1
dropbox~=11.36.2
xlrd~=2.0.1
pandas~=2.1.3
Unidecode~=1.3.7
matplotlib~=3.8.1
pyxlsb~=1.0.10
customtkinter~=5.2.1
Pillow~=10.1.0
openai~=1.22.0
bardapi~=0.1.38
SQLAlchemy~=2.0.23
reportlab
setuptools
keras~=3.3.3
tensorflow~=2.16.1
scikit-learn~=1.5.0
Office365-REST-Python-Client~=2.5.11
Werkzeug~=2.3.8
Flask-WTF
WTForms~=3.1.2
WTForms-JSON~=0.3.5
xlsxwriter
PyPDF2~=3.0.1
openpyxl
pyserial~=3.5
PyMuPDF